+++ types/modules/filters/index.d.ts
@@ -1,3 +1,0 @@
-import type Handler from "../handler.js";
-declare const handlers: Array<typeof Handler>;
-export default handlers;