+++ types/modules/generated-content/index.d.ts
@@ -1,3 +1,0 @@
-import type Handler from "../handler.js";
-declare const _default: Array<typeof Handler>;
-export default _default;